feat/water-2538 -- billing routes and controllers structure update -- tests updated for new folder structure
