+++ test/internal/modules/billing/controller.js
@@ -14,7 +14,7 @@ const uuid = require('uuid/v4');
 const { logger } = require('internal/logger');
 const forms = require('shared/lib/forms');
 const services = require('internal/lib/connectors/services');
-const controller = require('internal/modules/billing/controller');
+const controller = require('internal/modules/billing/controllers/bill-run');
 const batchService = require('internal/modules/billing/services/batch-service');
 const transactionsCSV = require('internal/modules/billing/services/transactions-csv');
 const csv = require('internal/lib/csv-download');
