+++ test/internal/modules/billing/routes.js
@@ -5,11 +5,12 @@ const { experiment, test } = exports.lab = require('@hapi/lab').script();
 
 const preHandlers = require('internal/modules/billing/pre-handlers');
 const { scope } = require('internal/lib/constants');
-const routes = require('internal/modules/billing/routes');
+const routes = require('internal/modules/billing/routes/bill-run');
 
 experiment('internal/modules/billing/routes', () => {
   experiment('.getBillingBatchSummary', () => {
     test('uses the redirectToWaitingIfEventNotCompleted pre handler', async () => {
+      console.log(routes);
       const routePreHandlers = routes.getBillingBatchSummary.config.pre;
       expect(routePreHandlers).to.contain(preHandlers.redirectToWaitingIfEventNotComplete);
     });
